Add ability to populate db with fake data
- `yarn db create-fake-data 20`
- `yarn db clean-fake-data`
