+++ package.json
@@ -47,6 +47,7 @@
     "@babel/core": "^7.12.10",
     "@babel/preset-env": "^7.12.11",
     "@babel/preset-typescript": "^7.12.7",
+    "@types/faker": "^5.1.5",
     "@types/jest": "^26.0.19",
     "@typescript-eslint/eslint-plugin": "^4.11.0",
     "@typescript-eslint/parser": "^4.11.0",
@@ -65,6 +66,7 @@
     "eslint-plugin-react-hooks": "^4.2.0",
     "eslint-plugin-simple-import-sort": "^7.0.0",
     "eslint_d": "^9.1.2",
+    "faker": "^5.1.0",
     "inquirer": "^7.3.3",
     "jest": "^26.6.3",
     "mock-req": "^0.2.0",
